python cipher e decipher a funcionar

+++ E-DES.c
@@ -157,8 +157,8 @@ void feistel_networks_block(SBox *SBoxes, uint8_t block[8], uint8_t output[8]) {
 
 // Loops through the blocks and encrypts them
 void feistel_networks(SBox *SBoxes, uint8_t *input, uint8_t *output, uint64_t input_len) {
-    int num_blocks = input_len / 8;
-    for (int i = 0; i < num_blocks; i++) {
+    uint64_t num_blocks = input_len / 8;
+    for (uint64_t i = 0; i < num_blocks; i++) {
         feistel_networks_block(SBoxes, &input[i * 8], &output[i * 8]);
     }
 }
@@ -216,8 +216,8 @@ void feistel_networks_block_decrypt(SBox *SBoxes, uint8_t block[8], uint8_t outp
 }
 
 void feistel_networks_decrypt(SBox *SBoxes, uint8_t *input, uint8_t *output, uint64_t input_len) {
-    int num_blocks = input_len / 8;
-    for (int i = 0; i < num_blocks; i++) {
+    uint64_t num_blocks = input_len / 8;
+    for (uint64_t i = 0; i < num_blocks; i++) {
         feistel_networks_block_decrypt(SBoxes, &input[i * 8], &output[i * 8]);
     }
 }
